#2 Editing Production Line Producer Attributes
There were a lot of changes made to the AppActions and Dispatcher.
When the dispatcher gets an action the payload is now different, it still contains the action and the data but the payload is completely passed to the action and the action must finish using the Actions.finish(payload) to emit changes on the desired stores and for each store there is a set of component ids. This may have been an unnecessary change but it could be useful when dealing with multiple components that need updating in a single store.