Backend reconstruction
Removed Utilities file and placed the code into the correct controllers. 
Created new function to update the production line when the speed of a producer is changed. Instead of morphing the number of assemblers to the output line, the items per second is updated to reflect the new amount based on the current number of assemblers